+++ energizer/__init__.py
@@ -1,5 +1,9 @@
 """Top-level package for Pytorch-Energizer."""
 
+from energizer.mixin.base import DeterministicMixin, MCDropoutMixin
+from energizer.trainer import Trainer
+
 __author__ = """Pietro Lesci"""
 __email__ = 'pietrolesci@outlook.com'
 __version__ = '0.1.0'
+__all__ = ["DeterministicMixin", "MCDropoutMixin", "Trainer"]
