+++ energizer/loops/pool_epoch_loop.py
@@ -5,8 +5,8 @@ from pytorch_lightning.loops.epoch.evaluation_epoch_loop import EvaluationEpochL
 from pytorch_lightning.utilities.model_helpers import is_overridden
 from pytorch_lightning.utilities.types import STEP_OUTPUT
 
-from energizer.learners.base import Learner
 from energizer.metrics import AccumulateTopK
+from energizer.mixin.base import Learner
 
 
 class PoolEvaluationEpochLoop(EvaluationEpochLoop):
